+++ src/Service/ProfilePhotoService.php
@@ -36,7 +36,6 @@ class ProfilePhotoService
 
             foreach ($profilePhotos as $profilePhoto) {
                 if ($type === $profilePhoto->getType()) {
-
                     $profilePhoto->deactivate();
                     $this->entityManager->persist($profilePhoto);
                 }
@@ -52,7 +51,6 @@ class ProfilePhotoService
             $this->entityManager->flush();
 
             return true;
-
         } catch (Exception $e) {
             return false;
         }
@@ -93,12 +91,12 @@ class ProfilePhotoService
 
             return [
                 'status' => 'success',
-                'photos' => $activePhotos
+                'photos' => $activePhotos,
             ];
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return [
                 'status' => 'error',
-                'message' => $e->getMessage()
+                'message' => $e->getMessage(),
             ];
         }
     }